+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/repair/DefaultRepairConfigurationProvider.java
@@ -22,13 +22,13 @@ import java.util.function.Function;
 import com.datastax.oss.driver.api.core.CqlIdentifier;
 import com.datastax.oss.driver.api.core.CqlSession;
 import com.datastax.oss.driver.api.core.metadata.Node;
+import com.datastax.oss.driver.api.core.metadata.NodeStateListenerBase;
 import com.datastax.oss.driver.api.core.metadata.schema.AggregateMetadata;
 import com.datastax.oss.driver.api.core.metadata.schema.FunctionMetadata;
 import com.datastax.oss.driver.api.core.metadata.schema.KeyspaceMetadata;
 import com.datastax.oss.driver.api.core.metadata.schema.SchemaChangeListener;
 import com.datastax.oss.driver.api.core.metadata.schema.TableMetadata;
 import com.datastax.oss.driver.api.core.metadata.schema.ViewMetadata;
-import com.datastax.oss.driver.api.core.metadata.NodeStateListener;
 import com.datastax.oss.driver.api.core.session.Session;
 import com.datastax.oss.driver.api.core.type.UserDefinedType;
 import com.ericsson.bss.cassandra.ecchronos.core.utils.Metadata;
@@ -36,13 +36,17 @@ import com.ericsson.bss.cassandra.ecchronos.core.utils.ReplicatedTableProvider;
 import com.ericsson.bss.cassandra.ecchronos.core.utils.TableReference;
 import com.ericsson.bss.cassandra.ecchronos.core.utils.TableReferenceFactory;
 import com.google.common.base.Preconditions;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
- * A repair configuration provider that adds configuration to {@link RepairScheduler} based on whether or not the table
+ * A repair configuration provider that adds configuration to {@link RepairScheduler} based on whether the table
  * is replicated locally using the default repair configuration provided during construction of this object.
  */
-public class DefaultRepairConfigurationProvider implements SchemaChangeListener, NodeStateListener // NOPMD
+public class DefaultRepairConfigurationProvider extends NodeStateListenerBase implements SchemaChangeListener
 {
+    private static final Logger LOG = LoggerFactory.getLogger(DefaultRepairConfigurationProvider.class);
+
     private CqlSession mySession;
     private ReplicatedTableProvider myReplicatedTableProvider;
     private RepairScheduler myRepairScheduler;
@@ -388,52 +392,30 @@ public class DefaultRepairConfigurationProvider implements SchemaChangeListener,
     public void onSessionReady(final Session session)
     {
         SchemaChangeListener.super.onSessionReady(session);
-        // prepare the NodeStateListener
-        NodeStateListener.super.onSessionReady(session);
-    }
-
-    /**
-     * Callback for when a node is added to the cluster.
-     *
-     * @param node
-     */
-    @Override
-    public void onAdd(final Node node)
-    {
-        // NOOP
     }
 
     /**
      * Callback for when a node switches state to UP.
      *
-     * @param node
+     * @param node The node switching state to UP
      */
     @Override
     public void onUp(final Node node)
     {
+        LOG.debug("{} switched state to UP.", node);
         setupConfiguration();
     }
 
     /**
      * Callback for when a node switches state to DOWN.
      *
-     * @param node
+     * @param node The node switching state to DOWN
      */
     @Override
     public void onDown(final Node node)
     {
+        LOG.debug("{} switched state to DOWN.", node);
         setupConfiguration();
-    }
-
-    /**
-     * Callback for when a node is removed from the cluster.
-     *
-     * @param node
-     */
-    @Override
-    public void onRemove(final Node node)
-    {
-        // NOOP
     }
 
     /**
@@ -444,6 +426,7 @@ public class DefaultRepairConfigurationProvider implements SchemaChangeListener,
     {
         if (mySession == null)
         {
+            LOG.debug("Session during setupConfiguration call was null.");
             return;
         }